+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Media Naming: Sync file naming configuration even if `rename` is not set to `true`.
 - Quality Profiles: Validation check added for quality groups with less than 2 qualities.
 - Quality Profiles: Fix "Groups must contain multiple qualities" sync error.
+- Quality Profiles: Fix "Must contain all qualities" sync error.
 
 ## [6.0.0] - 2023-09-29